# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tribe management system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scouts:
    get:
      operationId: listScouts
      tags: [scouts]
      summary: List all scouts
      parameters:
        - name: gizra
          in: query
          schema:
            type: ["string", "null"]
          description: Filter by gizra (kvutza group)
        - name: role
          in: query
          schema:
            type: ["string", "null"]
          description: Filter by role
      responses:
        "200":
          description: List of scouts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scout"
    post:
      operationId: createScout
      tags: [scouts]
      summary: Create a scout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScoutBody"
      responses:
        "201":
          description: Scout created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scout"
  /scouts/{id}:
    get:
      operationId: getScout
      tags: [scouts]
      summary: Get a scout by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scout"
        "404":
          description: Not found
    patch:
      operationId: updateScout
      tags: [scouts]
      summary: Update a scout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScoutBody"
      responses:
        "200":
          description: Updated scout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scout"
    delete:
      operationId: deleteScout
      tags: [scouts]
      summary: Delete a scout
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---- EVENTS (Mifaalim) ----
...
components:
  schemas:
...
    Scout:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        gizra:
          type: string
          description: "Kvutza group (e.g., gizra name)"
        role:
          type: string
          description: "Role in tribe: marcaz_vogger, marcaz_tzair, roshatz, roshgad, chanich"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - gizra
        - role
        - createdAt
    CreateScoutBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        gizra:
          type: string
        role:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - name
        - gizra
        - role
    UpdateScoutBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        gizra:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...